myBlog/requirements.txt
amqp==5.0.6
asgiref==3.4.1
attrs==21.2.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coverage==5.5
Django==3.2.5
django-celery-beat==2.2.1
django-pytest==0.2.0
django-redis==5.0.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
factory-boy==3.2.0
Faker==8.10.0
idna==2.10
iniconfig==1.1.1
kombu==5.1.0
model-bakery==1.3.2
packaging==21.0
pluggy==0.13.1
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
py==1.10.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.12.1
pytest-django==4.4.0
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
